# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AutoCare API — vehicle owners book services, service centers fulfill them.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings for the active role
      parameters:
        - name: role
          in: query
          required: false
          schema: { type: string, enum: [owner, center] }
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [requested, accepted, in_progress, awaiting_approval, approved, completed, cancelled] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Booking" }
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Owner books a service
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBookingInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      parameters:
        - { name: bookingId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookingDetail" }
  /bookings/{bookingId}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Move a booking through its lifecycle
      parameters:
        - { name: bookingId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBookingStatusInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
  /bookings/{bookingId}/assign-mechanic:
    post:
      operationId: assignMechanic
      tags: [bookings]
      summary: Service center assigns a mechanic to a booking
      parameters:
        - { name: bookingId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignMechanicInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
  /bookings/{bookingId}/invoice:
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Service center creates an invoice for a booking
      parameters:
        - { name: bookingId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInvoiceInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
  /invoices/{invoiceId}:
    get:
      operationId: getInvoice
      tags: [invoices]
      parameters:
        - { name: invoiceId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
  /invoices/{invoiceId}/approve:
    post:
      operationId: approveInvoice
      tags: [invoices]
      summary: Owner approves a pending invoice
      parameters:
        - { name: invoiceId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
  /invoices/{invoiceId}/pay:
    post:
      operationId: payInvoice
      tags: [invoices]
      summary: Mark an approved invoice as paid (mocked payment)
      parameters:
        - { name: invoiceId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
...
components:
  schemas:
...
    Vehicle:
      type: object
      required: [id, ownerName, brand, model, year, plateNumber, color, mileage, serviceIntervalDays, serviceIntervalKm, createdAt]
      properties:
        id: { type: string }
        ownerName: { type: string }
        ownerPhone: { type: string, nullable: true }
        brand: { type: string }
        model: { type: string }
        year: { type: integer }
        vin: { type: string, nullable: true }
        plateNumber: { type: string }
        color: { type: string }
        engineType: { type: string, nullable: true }
        mileage: { type: integer }
        imageUrl: { type: string, nullable: true }
        nextServiceDate: { type: string, format: date-time, nullable: true }
        insuranceProvider: { type: string, nullable: true }
        serviceIntervalDays: { type: integer, minimum: 1 }
        serviceIntervalKm: { type: integer, minimum: 1 }
        lastServicedAt: { type: string, format: date-time, nullable: true }
        lastServicedMileage: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
...
    ServiceCenter:
      type: object
      required: [id, name, address, phone, specialties, rating, reviewsCount, openJobs]
      properties:
        id: { type: string }
        name: { type: string }
        address: { type: string }
        city: { type: string }
        region: { type: string }
        phone: { type: string }
        specialties:
          type: array
          items: { type: string }
        rating: { type: number }
        reviewsCount: { type: integer }
        openJobs: { type: integer }
        imageUrl: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        active: { type: boolean }
        whatsappOptIn: { type: boolean }
    Mechanic:
      type: object
      required: [id, serviceCenterId, name, yearsExperience, specialization, rating, completedJobs]
      properties:
        id: { type: string }
        serviceCenterId: { type: string }
        name: { type: string }
        yearsExperience: { type: integer }
        specialization: { type: string }
        certifications:
          type: array
          items: { type: string }
        rating: { type: number }
        completedJobs: { type: integer }
        avatarUrl: { type: string, nullable: true }
        active: { type: boolean }
    Booking:
      type: object
      required: [id, vehicleId, serviceCenterId, serviceType, description, status, requestedAt]
      properties:
        id: { type: string }
        vehicleId: { type: string }
        serviceCenterId: { type: string }
        mechanicId: { type: string, nullable: true }
        serviceType: { type: string }
        description: { type: string }
        status:
          type: string
          enum: [requested, accepted, in_progress, awaiting_approval, approved, completed, cancelled]
        requestedAt: { type: string, format: date-time }
        scheduledAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        estimatedDurationHours: { type: number, nullable: true }
        vehicle: { $ref: "#/components/schemas/Vehicle" }
        serviceCenter: { $ref: "#/components/schemas/ServiceCenter" }
        mechanic:
          oneOf:
            - { $ref: "#/components/schemas/Mechanic" }
            - { type: "null" }
        invoiceId: { type: string, nullable: true }
    BookingDetail:
      allOf:
        - $ref: "#/components/schemas/Booking"
        - type: object
          properties:
            invoice:
              oneOf:
                - { $ref: "#/components/schemas/Invoice" }
                - { type: "null" }
            timeline:
              type: array
              items: { $ref: "#/components/schemas/TimelineEntry" }
    TimelineEntry:
      type: object
      required: [at, label]
      properties:
        at: { type: string, format: date-time }
        label: { type: string }
        actor: { type: string, nullable: true }
    CreateBookingInput:
      type: object
      required: [vehicleId, serviceCenterId, serviceType, description]
      properties:
        vehicleId: { type: string }
        serviceCenterId: { type: string }
        serviceType: { type: string }
        description: { type: string }
        scheduledAt: { type: string, format: date-time, nullable: true }
    UpdateBookingStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [requested, accepted, in_progress, awaiting_approval, approved, completed, cancelled]
        estimatedDurationHours: { type: number, nullable: true }
        scheduledAt: { type: string, format: date-time, nullable: true }
    AssignMechanicInput:
      type: object
      required: [mechanicId]
      properties:
        mechanicId: { type: string }
    InvoiceItem:
      type: object
      required: [description, quantity, unitPrice, kind]
      properties:
        description: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        kind:
          type: string
          enum: [labor, part]
    Invoice:
      type: object
      required: [id, bookingId, items, laborTotal, partsTotal, tax, total, status, createdAt]
      properties:
        id: { type: string }
        bookingId: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/InvoiceItem" }
        laborTotal: { type: number }
        partsTotal: { type: number }
        tax: { type: number }
        total: { type: number }
        notes: { type: string, nullable: true }
        status:
          type: string
          enum: [pending_approval, approved, paid, rejected]
        createdAt: { type: string, format: date-time }
        approvedAt: { type: string, format: date-time, nullable: true }
        paidAt: { type: string, format: date-time, nullable: true }
    CreateInvoiceInput:
      type: object
      required: [items, taxRate]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/InvoiceItem" }
        taxRate: { type: number, minimum: 0, maximum: 1 }
        notes: { type: string, nullable: true }
...