# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AutoCare API — vehicle owners book services, service centers fulfill them.
 * OpenAPI spec version: 0.1.0
 */

export interface Vehicle {
  id: string;
  ownerName: string;
  ownerPhone?: string | null;
  brand: string;
  model: string;
  year: number;
  vin?: string | null;
  plateNumber: string;
  color: string;
  engineType?: string | null;
  mileage: number;
  imageUrl?: string | null;
  nextServiceDate?: Date | null;
  insuranceProvider?: string | null;
  /** @minimum 1 */
  serviceIntervalDays: number;
  /** @minimum 1 */
  serviceIntervalKm: number;
  lastServicedAt?: Date | null;
  lastServicedMileage?: number | null;
  createdAt: Date;
}